Fix mixed-dtype sdpa for AMP/TPU: cast K/V to Q's dtype
F.scaled_dot_product_attention requires all inputs same dtype.
Under torch.autocast, Q (from backbone) gets promoted to bf16
but K/V (from canvas projections with f32 weights) stay f32.

Cast K/V to q.dtype before sdpa. No-op when all same dtype.
Enables torch.autocast('xla', dtype=torch.bfloat16) on TPU.

+++ canvit_pytorch/attention/base.py
@@ -79,5 +79,7 @@ class CanvasAttention(nn.Module):
         q = rope_apply_with_prefix(x=q, rope=query_rope)
         k = rope_apply_with_prefix(x=k, rope=kv_rope)
 
-        out: Tensor = F.scaled_dot_product_attention(q, k, v)
+        # Cast K/V to Q's dtype for AMP compatibility (e.g. TPU autocast
+        # promotes Q to bf16 via backbone but K/V stay f32 from canvas projections).
+        out: Tensor = F.scaled_dot_product_attention(q, k.to(q.dtype), v.to(q.dtype))
         return self.out_proj(from_multihead(out))
